requirements.txt
simple-salesforce>=1.12.6
python-dotenv>=1.0.0
requests>=2.31.0
PyYAML>=6.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
